Update New-Win32App.ps1
C:\ADOAgent\_work\1\s\Scripts\New-Win32App.ps1 : Cannot validate argument on parameter 'FilePath'. Cannot bind  argument to parameter 'Path' because it is an empty string.

"Temporary" workaround, makes it so that the intunewinapputil.exe is always found since it is in $env.
it is not a good long term solution but at least it gets everything working again.
I will need someone with a better grasp to fix this in a more long term manner.